refactor: refactor the package structure related to the Onboarding feature.
