+++ app/src/main/java/com/moscow/tudee/presentation/screen/onboarding/components/OnBoardingImage.kt
@@ -1,4 +1,4 @@
-package com.moscow.tudee.presentation.screen.onboarding
+package com.moscow.tudee.presentation.screen.onboarding.components
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.LinearOutSlowInEasing
@@ -20,6 +20,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
+import com.moscow.tudee.presentation.screen.onboarding.OnboardingData
 import kotlinx.coroutines.delay
 
 @Composable
